requirements.txt
python-dotenv
streamlit
streamlit-option-menu
matplotlib
numpy
firebase-admin
pandas
scikit-learn
python-dotenv
